seperate image pre-processing from ocr code
